+++ praktikumsplaner-frontend/frontend/src/components/assign/NwkCard.vue
